requirements.txt
Click==7.0
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
mongokit-py3==0.9.1.1
pymongo==2.9
Werkzeug==0.16.0
